:recycle: WIP fix account management bugs
