+++ src/modules/account/utils/index.js
@@ -1,2 +1,1 @@
 export * from './encryptAccount';
-export * from './decryptAccount';
